+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/gallery/AttachmentGalleryActivity.kt
@@ -22,18 +22,22 @@ import android.net.Uri
 import android.os.Bundle
 import android.os.Parcelable
 import android.text.format.DateUtils
+import android.view.View
+import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintSet
 import androidx.core.content.ContextCompat
 import androidx.core.view.isVisible
+import androidx.lifecycle.lifecycleScope
 import androidx.viewpager2.widget.ViewPager2
 import com.getstream.sdk.chat.StreamFileUtil
 import com.getstream.sdk.chat.images.StreamImageLoader
 import com.getstream.sdk.chat.utils.PermissionChecker
 import com.getstream.sdk.chat.utils.extensions.constrainViewToParentBySide
-import com.getstream.sdk.chat.utils.extensions.imagePreviewUrl
 import com.getstream.sdk.chat.utils.formatTime
+import io.getstream.chat.android.client.models.Attachment
+import io.getstream.chat.android.client.models.AttachmentType
 import io.getstream.chat.android.core.internal.coroutines.DispatcherProvider
 import io.getstream.chat.android.ui.ChatUI
 import io.getstream.chat.android.ui.R
@@ -48,9 +52,9 @@ import io.getstream.chat.android.ui.gallery.options.AttachmentGalleryOptionsView
 import io.getstream.chat.android.ui.gallery.options.internal.AttachmentGalleryOptionsDialogFragment
 import io.getstream.chat.android.ui.gallery.overview.internal.MediaAttachmentDialogFragment
 import io.getstream.logging.StreamLog
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.delay
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import kotlinx.parcelize.Parcelize
 import java.util.Date
 
@@ -71,14 +75,22 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
     private var showInChatOptionEnabled: Boolean = true
 
     /**
-     * If the "save image" option is present in the list
+     * If the "save media" option is present in the list
      */
-    private var saveImageOptionEnabled: Boolean = true
+    private var saveMediaOptionEnabled: Boolean = true
 
     /**
      * If the "delete" option is present in the list.
      */
     private var deleteOptionEnabled: Boolean = true
+
+    /**
+     * Represents the job used to share an attachment.
+     *
+     * If the attachment is larger, this could end up being a longer run job
+     * so cancel it accordingly.
+     */
+    private var shareMediaJob: Job? = null
 
     private val initialIndex: Int by lazy { intent.getIntExtra(EXTRA_KEY_INITIAL_INDEX, 0) }
     private val viewModel: AttachmentGalleryViewModel by viewModels()
@@ -98,21 +110,6 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
         }
     private var isFullScreen = false
 
-    private var onSharePictureListener: (pictureUri: Uri) -> Unit = { pictureUri ->
-        ContextCompat.startActivity(
-            this,
-            Intent.createChooser(
-                Intent(Intent.ACTION_SEND).apply {
-                    type = "image/*"
-                    putExtra(Intent.EXTRA_STREAM, pictureUri)
-                    addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-                },
-                getString(R.string.stream_ui_attachment_gallery_share),
-            ),
-            null
-        )
-    }
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
@@ -129,7 +126,7 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
         } else {
             this.attachmentGalleryItems = attachmentGalleryItems
             setupGalleryAdapter()
-            setupShareImageButton()
+            setupShareMediaButton()
             setupAttachmentActionsButton()
             obtainOptionsViewStyle()
             observePageChanges()
@@ -139,8 +136,8 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
     private fun setupGalleryAdapter() {
         adapter = AttachmentGalleryPagerAdapter(
             fragmentActivity = this,
-            imageList = attachmentGalleryItems.map { it.attachment.imagePreviewUrl!! },
-            imageClickListener = {
+            mediaList = attachmentGalleryItems.map { it.attachment },
+            mediaClickListener = {
                 isFullScreen = !isFullScreen
                 if (isFullScreen) enterFullScreenMode() else exitFullScreenMode()
             }
@@ -149,30 +146,140 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
         binding.galleryViewPager.setCurrentItem(initialIndex, false)
     }
 
-    private fun setupShareImageButton() {
-        binding.shareImageButton.setOnClickListener {
-            it.isEnabled = false
-            GlobalScope.launch(DispatcherProvider.Main) {
-                StreamImageLoader.instance().loadAsBitmap(
-                    context = applicationContext,
-                    url = adapter.getItem(binding.galleryViewPager.currentItem)
-                )?.let { bitmap ->
-                    StreamFileUtil.writeImageToSharableFile(applicationContext, bitmap)
-                }?.let(onSharePictureListener)
-
-                delay(500)
-                it.isEnabled = true
+    /**
+     * Sets an on click listener with media sharing capability
+     * on the share button.
+     */
+    private fun setupShareMediaButton() {
+        binding.shareMediaButton.setOnClickListener {
+            if (shareMediaJob == null || shareMediaJob?.isCompleted == true) {
+                setSharingInProgressUi()
+                shareMedia()
+            } else {
+                shareMediaJob?.cancel()
+                setNoSharingInProgressUi()
             }
         }
+    }
+
+    /**
+     * Begins the process of sharing media depending on
+     * the attachment type.
+     */
+    private fun shareMedia() {
+        val attachment = adapter.getItem(binding.galleryViewPager.currentItem)
+
+        when (attachment.type) {
+            AttachmentType.IMAGE -> shareImage(attachment = attachment)
+            AttachmentType.VIDEO -> shareVideo(attachment = attachment)
+            else -> toastFailedShare()
+        }
+    }
+
+    /**
+     * Prepares an image attachment for sharing.
+     *
+     * @param attachment The attachment to share.
+     **/
+    private fun shareImage(attachment: Attachment) {
+        val imageUrl = attachment.imageUrl
+
+        if (imageUrl != null) {
+            shareMediaJob?.cancel()
+
+            shareMediaJob = lifecycleScope.launch(DispatcherProvider.Main) {
+                StreamImageLoader.instance().loadAsBitmap(
+                    context = applicationContext,
+                    url = imageUrl
+                )?.let { bitmap ->
+                    StreamFileUtil.writeImageToSharableFile(applicationContext, bitmap)
+                }?.let {
+                    launchShareActivity(mediaUri = it, attachmentType = attachment.type)
+                }
+                setNoSharingInProgressUi()
+            }
+        } else {
+            toastFailedShare()
+            setNoSharingInProgressUi()
+        }
+    }
+
+    /**
+     * Prepares a video attachment for sharing.
+     *
+     * @param attachment The attachment to share.
+     */
+    private fun shareVideo(attachment: Attachment) {
+        shareMediaJob?.cancel()
+
+        shareMediaJob = lifecycleScope.launch(DispatcherProvider.IO) {
+            val result = withContext(DispatcherProvider.IO) {
+                StreamFileUtil.writeFileToShareableFile(
+                    context = applicationContext,
+                    attachment = attachment
+                )
+            }
+
+            if (result.isSuccess) {
+                launchShareActivity(mediaUri = result.data(), attachmentType = attachment.type)
+            } else {
+                withContext(DispatcherProvider.Main) {
+                    toastFailedShare()
+                }
+            }
+            withContext(DispatcherProvider.Main) {
+                setNoSharingInProgressUi()
+            }
+        }
+    }
+
+    private fun launchShareActivity(
+        mediaUri: Uri,
+        attachmentType: String?,
+    ) {
+        val type = when (attachmentType) {
+            AttachmentType.IMAGE -> "image/*"
+            AttachmentType.VIDEO -> "video/*"
+            else -> {
+                toastFailedShare()
+                return
+            }
+        }
+
+        ContextCompat.startActivity(
+            this,
+            Intent.createChooser(
+                Intent(Intent.ACTION_SEND).apply {
+                    this.type = type
+                    putExtra(Intent.EXTRA_STREAM, mediaUri)
+                    addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
+                },
+                getString(R.string.stream_ui_attachment_gallery_share),
+            ),
+            null
+        )
+    }
+
+    /**
+     * Displays a toast saying that sharing the attachment has failed.
+     */
+    private fun toastFailedShare() {
+        Toast.makeText(
+            applicationContext,
+            applicationContext.getString(R.string.stream_ui_attachment_gallery_share_error),
+            Toast.LENGTH_SHORT
+        ).show()
     }
 
     private fun setupAttachmentActionsButton() {
         binding.attachmentActionsButton.setOnClickListener {
             AttachmentGalleryOptionsDialogFragment.newInstance(
-                showInChatOptionHandler = { setResultAndFinish(AttachmentOptionResult.ShowInChat(attachmentGalleryResultItem)) },
+                showInChatOptionHandler = {
+                    setResultAndFinish(AttachmentOptionResult.ShowInChat(attachmentGalleryResultItem))
+                },
                 deleteOptionHandler = { setResultAndFinish(AttachmentOptionResult.Delete(attachmentGalleryResultItem)) },
                 replyOptionHandler = { setResultAndFinish(AttachmentOptionResult.Reply(attachmentGalleryResultItem)) },
-                saveImageOptionHandler = handleSaveImage,
+                saveMediaOptionHandler = handleSaveMedia,
                 isMessageMine = attachmentGalleryItems[binding.galleryViewPager.currentItem].isMine,
             ).show(supportFragmentManager, AttachmentGalleryOptionsDialogFragment.TAG)
         }
@@ -189,7 +296,7 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
         onGalleryPageSelected(initialIndex)
     }
 
-    private val handleSaveImage = AttachmentGalleryOptionsDialogFragment.AttachmentOptionHandler {
+    private val handleSaveMedia = AttachmentGalleryOptionsDialogFragment.AttachmentOptionHandler {
         permissionChecker.checkWriteStoragePermissions(
             binding.root,
             onPermissionGranted = {
@@ -207,7 +314,7 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
     }
 
     private fun onGalleryPageSelected(position: Int) {
-        binding.imageCountTextView.text = getString(
+        binding.mediaInformationTextView.text = getString(
             R.string.stream_ui_attachment_gallery_count,
             position + 1,
             attachmentGalleryItems.size
@@ -241,7 +348,7 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
         binding.galleryOverviewButton.setOnClickListener {
             MediaAttachmentDialogFragment.newInstance()
                 .apply {
-                    setImageClickListener {
+                    setMediaClickListener {
                         binding.galleryViewPager.setCurrentItem(it, true)
                         dismiss()
                     }
@@ -279,7 +386,47 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
     }
 
     /**
-     * Obtains style attributes for the options list.
+     * Sets up the UI in a way that signals to the user
+     * that sharing is in progress.
+     */
+    private fun setSharingInProgressUi() {
+        with(binding) {
+            mediaInformationTextView.text =
+                applicationContext.getString(R.string.stream_ui_attachment_gallery_preview_preparing)
+            progressBar.visibility = View.VISIBLE
+
+            val drawable =
+                ContextCompat.getDrawable(applicationContext, R.drawable.stream_ui_ic_clear)?.mutate()
+                    ?.apply { setTint(ContextCompat.getColor(applicationContext, R.color.stream_ui_black)) }
+
+            binding.shareMediaButton.setImageDrawable(drawable)
+        }
+    }
+
+    /**
+     * Sets up the UI in a way that signals to the user
+     * that no attachment is being shared currently.
+     */
+    private fun setNoSharingInProgressUi() {
+        with(binding) {
+            progressBar.visibility = View.GONE
+            shareMediaButton.setImageDrawable(
+                ContextCompat.getDrawable(
+                    applicationContext,
+                    R.drawable.stream_ui_ic_share
+                )
+            )
+
+            mediaInformationTextView.text = getString(
+                R.string.stream_ui_attachment_gallery_count,
+                galleryViewPager.currentItem + 1,
+                attachmentGalleryItems.size
+            )
+        }
+    }
+
+    /**
+     * Obtains style attributes for the gallery and its children.
      */
     private fun obtainOptionsViewStyle() {
         try {
@@ -292,11 +439,11 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
                 val style = AttachmentGalleryOptionsViewStyle(this, it)
                 replyOptionEnabled = style.replyOptionEnabled
                 showInChatOptionEnabled = style.showInChatOptionEnabled
-                saveImageOptionEnabled = style.saveImageOptionEnabled
+                saveMediaOptionEnabled = style.saveMediaOptionEnabled
                 deleteOptionEnabled = style.deleteOptionEnabled
             }
         } catch (e: Exception) {
-            logger.e(e) { "Failed to obtain style attribute for the options menu" }
+            logger.e(e) { "Failed to obtain styles" }
         }
     }
 
@@ -309,8 +456,17 @@ public class AttachmentGalleryActivity : AppCompatActivity() {
     private fun shouldShowOptionsButton(isMine: Boolean): Boolean {
         return replyOptionEnabled ||
             showInChatOptionEnabled ||
-            saveImageOptionEnabled ||
+            saveMediaOptionEnabled ||
             (deleteOptionEnabled && isMine)
+    }
+
+    /**
+     * Clear the cache when the activity is
+     * destroyed.
+     */
+    override fun onDestroy() {
+        super.onDestroy()
+        StreamFileUtil.clearStreamCache(context = applicationContext)
     }
 
     public fun interface AttachmentShowInChatOptionHandler {