+++ stream-chat-android-ui-components/src/test/kotlin/io/getstream/chat/android/ui/MessageListItemLiveDataTest.kt
@@ -79,6 +79,7 @@ internal class MessageListItemLiveDataTest {
         val typing: LiveData<List<User>> = MutableLiveData(listOf())
         val deletedMessageVisibility = MutableLiveData(DeletedMessageVisibility.ALWAYS_VISIBLE)
         val messageFooterVisibility = MutableLiveData<MessageFooterVisibility>(MessageFooterVisibility.LastInGroup)
+        val endOfNewMessages = MutableLiveData(true)
 
         return MessageListItemLiveData(
             currentUser = currentUser,
@@ -90,6 +91,7 @@ internal class MessageListItemLiveDataTest {
             deletedMessageVisibility = deletedMessageVisibility,
             messageFooterVisibility = messageFooterVisibility,
             messagePositionHandlerProvider = MessageListViewModel.MessagePositionHandler::defaultHandler,
+            endOfNewMessages = endOfNewMessages
         )
     }
 
@@ -99,6 +101,7 @@ internal class MessageListItemLiveDataTest {
         val typing: LiveData<List<User>> = MutableLiveData(listOf())
         val deletedMessageVisibility = MutableLiveData(DeletedMessageVisibility.ALWAYS_VISIBLE)
         val messageFooterVisibility = MutableLiveData<MessageFooterVisibility>(MessageFooterVisibility.LastInGroup)
+        val endOfNewMessages = MutableLiveData(true)
 
         return MessageListItemLiveData(
             currentUser = currentUser,
@@ -110,6 +113,7 @@ internal class MessageListItemLiveDataTest {
             deletedMessageVisibility = deletedMessageVisibility,
             messageFooterVisibility = messageFooterVisibility,
             messagePositionHandlerProvider = MessageListViewModel.MessagePositionHandler::defaultHandler,
+            endOfNewMessages = endOfNewMessages
         )
     }
 
@@ -134,6 +138,7 @@ internal class MessageListItemLiveDataTest {
         val typing: LiveData<List<User>> = MutableLiveData(listOf())
         val deletedMessageVisibility = MutableLiveData(DeletedMessageVisibility.ALWAYS_VISIBLE)
         val messageFooterVisibility = MutableLiveData<MessageFooterVisibility>(MessageFooterVisibility.LastInGroup)
+        val endOfNewMessages = MutableLiveData(true)
 
         return MessageListItemLiveData(
             currentUser = currentUser,
@@ -145,6 +150,7 @@ internal class MessageListItemLiveDataTest {
             deletedMessageVisibility = deletedMessageVisibility,
             messageFooterVisibility = messageFooterVisibility,
             messagePositionHandlerProvider = MessageListViewModel.MessagePositionHandler::defaultHandler,
+            endOfNewMessages = endOfNewMessages
         )
     }
 
@@ -255,7 +261,7 @@ internal class MessageListItemLiveDataTest {
         val testObserver: Observer<MessageListItemWrapper> = mock()
         messageListItemLd.observeForever(testObserver)
         messageListItemLd.typingChanged(listOf(currentUser.value!!))
-        verify(testObserver, times(4)).onChanged(any())
+        verify(testObserver, times(5)).onChanged(any())
     }
 
     @Test