+++ stream-chat-android-ui-utils/src/main/kotlin/io/getstream/chat/android/uiutils/extension/Channel.kt
@@ -34,7 +34,11 @@ import io.getstream.chat.android.client.utils.message.isSystem
  * @return Last message from the channel or null if it doesn't exist.
  */
 public fun Channel.getPreviewMessage(currentUser: User?): Message? =
-    messages.asSequence()
+    if (isInsideSearch) {
+        cachedLatestMessages
+    } else {
+        messages
+    }.asSequence()
         .filter { it.createdAt != null || it.createdLocallyAt != null }
         .filter { it.deletedAt == null }
         .filter { !it.silent }